Started deletePlace, deleteBlog, deleteService server acrtions. Added validation to deletePrice server action.

+++ action/deletePrice.ts
@@ -2,12 +2,19 @@
 
 import { handleMongooseError } from "@/lib/mongo"
 import Price from "@/models/Price";
+import { deleteSchema } from "@/schema/schema";
 import { updateTag } from "next/cache";
 
 
 
 export const deletePrice = async ( _id: string) => {
     try {
+         const res = deleteSchema.safeParse(_id);
+        if(res.error?.issues) {
+            console.log(res.error.issues)
+            return {failed: res.error.issues.map((item) => item.message)}
+        }
+        
         await Price.findByIdAndDelete(_id);
 
         updateTag("price-data");
